style(PalleteSettings): change inline classNames for tw styled component
